requirements.txt
flask==0.11.1
GoogleAppEngineMapReduce
